Dark - Light UI Transition
The transition between the light and the dark colors has been synced with the transition of the toggle switch. When the toggle switch is moving towards the dark mode, giving an effect that nightfall is coming over the application and similarly when the toggle switch is moving towards the light mode, an effect that daylight is coming over the application is given.